+++ libs/angular-modal-gallery/src/services/keyboard.service.spec.ts
@@ -35,46 +35,88 @@ function setupRouter(injector: KeyboardServiceConfig) {
 }
 
 describe('KeyboardService', () => {
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        providers: [
-          {
-            provide: KeyboardService,
-            useFactory: setupRouter,
-            deps: [KEYBOARD_CONFIGURATION]
-          },
-          {
-            provide: KEYBOARD_CONFIGURATION,
-            useValue: {}
-          }
-        ]
-      });
-    })
-  );
 
-  it(
-    'should instantiate service when inject service',
-    inject([KeyboardService], (service: KeyboardService) => {
-      expect(service instanceof KeyboardService).toEqual(true);
-    })
-  );
-
-  describe('#reset()', () => {
-    it(
-      'should call reset',
-      inject([KeyboardService], (service: KeyboardService) => {
-        service.reset();
+  describe('KeyboardService WITHOUT external config via forRoot()', () => {
+    beforeEach(
+      async(() => {
+        TestBed.configureTestingModule({
+          providers: [
+            {
+              provide: KeyboardService,
+              useFactory: setupRouter,
+              deps: [KEYBOARD_CONFIGURATION]
+            },
+            {
+              provide: KEYBOARD_CONFIGURATION,
+              useValue: {}
+            }
+          ]
+        });
       })
     );
+
+    it('should instantiate service when inject service',
+      inject([KeyboardService], (service: KeyboardService) => {
+        expect(service instanceof KeyboardService).toEqual(true);
+      })
+    );
+
+    describe('#reset()', () => {
+      it('should call reset',
+        inject([KeyboardService], (service: KeyboardService) => {
+          service.reset();
+        })
+      );
+    });
+
+    describe('#add()', () => {
+      it('should call add',
+        inject([KeyboardService], (service: KeyboardService) => {
+          service.add(() => null);
+        })
+      );
+    });
   });
 
-  describe('#add()', () => {
-    it(
-      'should call add',
-      inject([KeyboardService], (service: KeyboardService) => {
-        service.add(() => null);
+  describe('KeyboardService WITH external config via forRoot()', () => {
+    beforeEach(
+      async(() => {
+        TestBed.configureTestingModule({
+          providers: [
+            {
+              provide: KeyboardService,
+              useFactory: setupRouter,
+              deps: [KEYBOARD_CONFIGURATION]
+            },
+            {
+              provide: KEYBOARD_CONFIGURATION,
+              useValue: {shortcuts: ['ctrl+a', 'ctrl+s']}
+            }
+          ]
+        });
       })
     );
+
+    it('should instantiate service when inject service',
+      inject([KeyboardService], (service: KeyboardService) => {
+        expect(service instanceof KeyboardService).toEqual(true);
+      })
+    );
+
+    describe('#reset()', () => {
+      it('should call reset',
+        inject([KeyboardService], (service: KeyboardService) => {
+          service.reset();
+        })
+      );
+    });
+
+    describe('#add()', () => {
+      it('should call add',
+        inject([KeyboardService], (service: KeyboardService) => {
+          service.add(() => null);
+        })
+      );
+    });
   });
 });